Refactor Assoc
