+++ lib/Dialogflow.ts
@@ -1,6 +1,5 @@
 import { IHttp, IHttpRequest, IModify, IPersistence, IRead } from '@rocket.chat/apps-engine/definition/accessors';
 import { ILivechatRoom } from '@rocket.chat/apps-engine/definition/livechat/ILivechatRoom';
-import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
 import { IRoom } from '@rocket.chat/apps-engine/definition/rooms';
 import { createSign } from 'crypto';
 import { AppSetting } from '../config/Settings';
@@ -9,7 +8,7 @@ import { Headers } from '../enum/Http';
 import { Logs } from '../enum/Logs';
 import { base64urlEncode } from './Helper';
 import { createHttpRequest } from './Http';
-import { retrieveDataByAssociation } from './retrieveDataByAssociation';
+import { retrieveDataByAssociation, RoomAssoc } from './Persistence';
 import { updateRoomCustomFields } from './Room';
 import { getAppSettingValue } from './Settings';
 
@@ -27,8 +26,7 @@ class DialogflowClass {
 
         const serverURL = await this.getServerURL(read, modify, http, sessionId);
 
-        const assoc = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `SFLAIA-${sessionId}`);
-        const data = await retrieveDataByAssociation(read, assoc);
+        const data = await retrieveDataByAssociation(read, RoomAssoc(sessionId));
 
         const defaultLanguageCode = await getAppSettingValue(read, AppSetting.DialogflowDefaultLanguage);
