+++ lib/Persistence.ts
@@ -1,5 +1,7 @@
 import { IRead } from '@rocket.chat/apps-engine/definition/accessors';
-import { RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
+import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
+
+export const RoomAssoc = (rid: string) => new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `DF-RID-${rid}`);
 
 export const retrieveDataByAssociation = async (read: IRead,  assoc: RocketChatAssociationRecord) => {
 
@@ -10,5 +12,4 @@ export const retrieveDataByAssociation = async (read: IRead,  assoc: RocketChatA
     }
 
     return {};
-
 };
