+++ lib/sessionMaintenance/SessionMaintenanceProcessor.ts
@@ -1,12 +1,11 @@
 import { IHttp, IModify, IPersistence, IRead } from '@rocket.chat/apps-engine/definition/accessors';
 import { ILivechatRoom } from '@rocket.chat/apps-engine/definition/livechat';
-import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
 import { IJobContext, IProcessor } from '@rocket.chat/apps-engine/definition/scheduler';
 import { AppSetting } from '../../config/Settings';
 import { DialogflowRequestType, LanguageCode } from '../../enum/Dialogflow';
 import { Dialogflow } from '../../lib/Dialogflow';
+import { retrieveDataByAssociation, RoomAssoc } from '../../lib/Persistence';
 import { getAppSettingValue } from '../../lib/Settings';
-import { retrieveDataByAssociation } from '../retrieveDataByAssociation';
 import { SessionMaintenanceOnceSchedule } from './SessionMaintenanceOnceSchedule';
 
 export class SessionMaintenanceProcessor implements IProcessor {
@@ -36,8 +35,7 @@ export class SessionMaintenanceProcessor implements IProcessor {
             return;
         }
 
-        const assoc = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, `SFLAIA-${jobContext.sessionId}`);
-        const data = await retrieveDataByAssociation(read, assoc);
+        const data = await retrieveDataByAssociation(read, RoomAssoc(jobContext.sessionId));
 
         const defaultLanguageCode = await getAppSettingValue(read, AppSetting.DialogflowDefaultLanguage);
